refactor: split standardizer and update the composition entry flow
